Per review, move logic to display ToS banner to a helper
It easier to understand when we can see the logic to display the banner
in the view.
